Extract SendAmountMixin
