+++ src/pages/mobile/SendAmount.vue
@@ -71,8 +71,7 @@
 <script>
 import { pick } from 'lodash-es';
 import { mapGetters } from 'vuex';
-import BigNumber from 'bignumber.js';
-import { TxBuilder } from '@aeternity/aepp-sdk/es';
+import SendAmountMixin from '../SendAmountMixin';
 import MobilePage from '../../components/mobile/Page.vue';
 import Guide from '../../components/Guide.vue';
 import AeFraction from '../../components/AeFraction.vue';
@@ -94,6 +93,7 @@ export default {
     DetailsAmountCurrency,
     AeButton,
   },
+  mixins: [SendAmountMixin],
   props: {
     to: {
       type: String,
@@ -101,55 +101,11 @@ export default {
     },
   },
   data: () => ({
-    customAmount: '',
-    maxSelected: false,
-    minFee: BigNumber(0),
     MAGNITUDE,
   }),
-  computed: {
-    ...mapGetters('accounts', ['activeColor']),
-    max() {
-      const max = this.activeAccount.balance.minus(this.minFee);
-      return (max.isPositive() ? max : 0).toString();
-    },
-    amount: {
-      get() {
-        return this.maxSelected ? this.max : this.customAmount;
-      },
-      set(value) {
-        if (value === this.max) this.maxSelected = true;
-        else {
-          this.customAmount = value;
-          this.maxSelected = false;
-        }
-      },
-    },
-  },
+  computed: mapGetters('accounts', ['activeColor']),
   subscriptions() {
     return pick(this.$store.state.observables, ['activeAccount']);
-  },
-  mounted() {
-    this.$watch(
-      ({ activeAccount: { address, nonce }, amount }) => ({ address, nonce, amount }),
-      ({ address, nonce, amount }) => {
-        const minFee = BigNumber(TxBuilder.calculateMinFee(
-          'spendTx', {
-            gas: this.$store.state.sdk.Ae.defaults.gas,
-            params: {
-              ...this.$store.state.sdk.Ae.defaults,
-              senderId: address,
-              recipientId: address,
-              amount: BigNumber(amount > 0 ? amount : 0).shiftedBy(MAGNITUDE),
-              ttl: 0,
-              nonce: nonce + 1,
-              payload: '',
-            },
-          },
-        )).shiftedBy(-MAGNITUDE);
-        if (!minFee.isEqualTo(this.minFee)) this.minFee = minFee;
-      },
-      { immediate: true },
-    );
   },
   methods: {
     async setAmount() {
